Add files via upload
This E-Commerce Application in Android is also very useful if you own a small shop business where you want to sell your products online. The main features of E-Commerce Application in Android are to view all products, choose products by category, search products, and view product details. By purchasing in the E-Commerce Application in Android you have to choose a product, and then add it in to cart. Indicate your preferred quantity then select checkout if you want to purchase the products. The E-Commerce Application in Android runs smoothly on your device as long as there’s an internet connection and it is more efficient and faster.